+++ linkerd/app/core/Cargo.toml
@@ -15,6 +15,7 @@ independently of the inbound and outbound proxy logic.
 [dependencies]
 bytes = "1"
 http = "0.2"
+html-escape = "0.2"
 http-body = "0.4"
 hyper = { version = "0.14.2", features = ["http1", "http2"] }
 futures = "0.3.15"
@@ -35,7 +36,6 @@ linkerd-error-respond = { path = "../../error-respond" }
 linkerd-exp-backoff = { path = "../../exp-backoff" }
 linkerd-http-classify = { path = "../../http-classify" }
 linkerd-http-metrics = { path = "../../http-metrics" }
-
 linkerd-http-retry = { path = "../../http-retry" }
 linkerd-identity = { path = "../../identity" }
 linkerd-io = { path = "../../io" }
@@ -63,6 +63,7 @@ linkerd-stack-tracing = { path = "../../stack/tracing" }
 linkerd-tls = { path = "../../tls" }
 linkerd-trace-context = { path = "../../trace-context" }
 regex = "1.0.0"
+serde_json = "1"
 thiserror = "1.0"
 tokio = { version = "1", features = ["macros", "sync", "parking_lot"]}
 tokio-stream = { version = "0.1.5", features = ["time"] }
