+++ linkerd/tracing/Cargo.toml
@@ -11,12 +11,7 @@ default = []
 ansi = ["tracing-subscriber/ansi"]
 
 [dependencies]
-bytes = "1"
-html-escape = "0.2.5"
-http = "0.2"
-hyper = { version = "0.14.2", features = ["http1"] }
 linkerd-error = { path = "../error" }
-serde_json = "1"
 tokio = { version = "1", features = ["time"] }
 tokio-trace = { git = "https://github.com/hawkw/tokio-trace", rev = "7d5998e7cb3beb06ada5983675319dc4853576c5", features = ["serde"] }
 tracing = "0.1.23"
